Update separating core from handlers.
